# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gantt App API",
    "description": "API REST para gestión de proyectos y tareas con diagramas Gantt",
    "contact": {
      "name": "Gantt App Team"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base URL"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "description": "Valida credenciales y devuelve un token Sanctum",
        "operationId": "99d66635c4992aeaa6aa44ff653d0563",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "admin@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  },
                  "device_name": {
                    "type": "string",
                    "example": "postman"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "token_type": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object"
                        }
                      },
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación"
          },
          "429": {
            "description": "Rate limit excedido"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Cerrar sesión",
        "description": "Revoca el token actual",
        "operationId": "6abf3b64a4bc7838d56346f05a5153af",
        "responses": {
          "200": {
            "description": "Sesión cerrada"
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Cerrar todas las sesiones",
        "description": "Revoca todos los tokens del usuario",
        "operationId": "8a0e47e08fb7ff4a767e76fd65943833",
        "responses": {
          "200": {
            "description": "Sesiones cerradas"
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Obtener usuario actual",
        "description": "Devuelve el usuario autenticado",
        "operationId": "7edae2cdbaafc10dff8c9c4198b320a2",
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "description": "Token de autenticación Laravel Sanctum",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}